Rename "tables.*" to "schemas.*" and group together
