+++ package.json
@@ -53,6 +53,8 @@
     "mysql2": "^1.4.0",
     "ng-inline-svg": "^5.0.0",
     "normalize.css": "^7.0.0",
+    "passport": "^0.4.0",
+    "passport-local": "^1.0.0",
     "reflect-metadata": "^0.1.10",
     "rxjs": "^5.4.2",
     "snyk": "^1.47.0",
@@ -73,6 +75,8 @@
     "@types/morgan": "^1.7.32",
     "@types/mysql2": "types/mysql2",
     "@types/node": "^8.0.22",
+    "@types/passport": "^0.4.1",
+    "@types/passport-local": "^1.0.32",
     "@types/sinon": "^4.0.0",
     "@types/sinon-chai": "^2.7.28",
     "@types/supertest": "^2.0.2",
